+++ packages/core/build.sh
@@ -3,4 +3,3 @@ npm install
 
 echo "building typescript..."
 npx tsc
-chmod +x dist/bin/index.js
