--- src/index.ts
@@ -1,3 +1,0 @@
-export {TartanExport} from "./tartan-export.js";
-export {TartanConfig, tartanConfigSchema} from "./tartan-config.js";
-export {TartanContext, tartanContextSchema} from "./tartan-context.js";
